Extract shared source routing
