+++ src/playback.ts
@@ -19,7 +19,6 @@
  */
 
 import { BasePlayback } from "./basePlayback";
-import type { Bus } from "./bus";
 import type { BaseSound, FadeType, LoopCount, PanType } from "./cacophony";
 import type {
   AudioBuffer,
@@ -59,18 +58,6 @@ export class Playback extends BasePlayback implements BaseSound {
   _fadeInConfig?: { duration: number; type: FadeType; perLoop: boolean; targetVolume: number };
   _fadeOutConfig?: { duration: number; type: FadeType };
   _loopEndCallback?: () => void;
-  /**
-   * Per-playback send-gain allocations: bus → allocated GainNode. Sounds
-   * record send intent in `Sound._sends` (value-only), and the actual
-   * GainNode lifecycle is owned here — allocated at preplay or
-   * `Sound.routeTo(bus, gain)`, torn down in {@link cleanup}.
-   *
-   * Iterable Map (not WeakMap) so cleanup can walk every send and call
-   * `disconnect()` on it; relying on GC for disconnect would leave the bus
-   * graph holding the allocation indirectly.
-   */
-  _sendGains: Map<Bus, GainNode> = new Map();
-
   /**
    * Desired pitch-shift factor (1 = no shift, 2 = +1 octave, 0.5 = -1 octave).
    * Stored even before the worklet node exists so a value set on a cleaned-up /
@@ -583,17 +570,6 @@ export class Playback extends BasePlayback implements BaseSound {
     this.currentLoop = 0;
     this.source.disconnect();
     this.source = undefined;
-    // Tear down any per-playback send-gain nodes allocated by Sound.routeTo
-    // (or preplay) before the gainNode is severed by super.cleanup(). The
-    // gainNode → sendGain → bus.input chain is now fully disconnected.
-    for (const sendGain of this._sendGains.values()) {
-      try {
-        sendGain.disconnect();
-      } catch {
-        // Best-effort — node may already have been disconnected externally.
-      }
-    }
-    this._sendGains.clear();
     // Tear down the phase-vocoder pitch-shift worklet node if one was spliced in.
     if (this._pitchShiftNode) {
       try {
